Working on bootloader (#9)
Improving the stage1 of the bootloader and making possible to boot using also a hard drive, after fixing a nasty problem in the build system. (exos.img was less than 2880 sectors, due to a previous unlucky change in the main Makefile.)

The exos.img file has been tested on a real PC by booting using an usbstick (seen as drive 0), as well.