Context switch (#14)
Implementing the actual context switch function and discarding the older switch to user mode one.
Now, even if there still no "processes", the schedule() function uses context_switch() to return to the process, instead of using the regular return code. This means that it will be enough to save 'regs' to a process entry in order to switch among many processes.