Actually handling nested interrupts and introducing kernel threads (#26)
This huge change actually makes nested interrupts to be handled correctly (hopefully!) and introduces kernel threads, with all the special code for handling them. Along with that, tons of small bugs have been fixed and many improvements have been added as well as a side-effect of the whole work, like the panic handling and dump of backtraces that now cannot cause page faults since checks with is_mapped() each pointer before de-referencing it.

Kernel tasklets have been just introduced as well, along with unit tests which, as a side-effect, improved the whole testing infrastructure as well.

Final note: this change fixes also a nasty bug preventing the bootloader to actually load stage2+kernel on real hardware due to a wrong number of cylinders in the fatpart. See the change in its script.