Introducing kernel preemption for user processes (#27)
This change makes each user thread to have its own kernel stack, that will be used by the kernel when an interrupt (or syscall) occurs. That code runs by default with preemption disabled but it can be enabled as well, as happens in the DEMO implementation of sys_waitpid(). Therefore, the kernel can actually preempt itself and correctly account the ticks that it spent working on a job for a user process. In addition to that, a kernel_yield() function has been added. That function allows kernel code to save its current state and call the scheduler. Overall, all that brings exOS much closer to modern operating systems like Linux.


Single commits
---------------------------------------------
* initial change

* getting rid of jiffies_from_switch

* fixing the accounting logic

* minor changes

* adding running_in_kernel and using it

* adding running_in_kernel and using it

* making each user thread to have a kernel stack

* waitpid with kernel preemption cannot work because of nested interrupts checks. That has to be re-engineered.

* kernel working for user app can be preempted correctly!

* making printk to disable preemption

* starting working on kernel state save

* implementing kernel_yield

* minor stuff

* make the main() to just call schedule()

* fixing finally the problem triggered by kfree() in remove task

* minor fixes. Now it is possible to count kernel ticks