Introducing partial support for FAT16 and FAT32 (#42)
This change introduces in the kernel a partial read-only support for FAT16 and
FAT32. The kernel now it is able to walk through the directory-tree of a FAT
file-system and read a whole file. The walk-through across cluster chains is
fully supported, both for directory entries and files. Only the "Long Directory
Entries" feature has not been implemented yet, but it certainly will. Since this
will be the FS used by the system, the 11 chars length-limit is not acceptable.

Intentionally not implemented: handling of bad clusters and fall-back to other
FATs different than the first one. That's clearly because this code has been
designed to work only with FAT volumes in ramdisks, where bad clusters simply do
not exist.