Introducing an EFI bootloader (#43)
This change introduces a EFI bootloader in the project. The generated image now
has both the custom BIOS bootloader embedded in the MBR and an EFI loader in
/EFI/BOOT/ (only for 64-bit machines). Limitations: unfortunately, I did not
find a way to set the video card in text mode using the EFI APIs so I had to go
back to real mode in switchmode.bin and execute int 0x10 for that, before
booting the kernel. The problem is that this approach works only if the firmware
is set to support also legacy bootloaders. In a pure-EFI environment, my code
will go back to real mode, but int 0x10 just won't work.

Therefore, the above limitation will remain for a long time unless:
- I find a way to set the video card in text mode without int 0x10
- I implement a graphical framebuffer in the kernel