[kmalloc/pageframe_allocator] Add alloc-8 and make virtual_* to use it
This change introduces alloc_8_pageframes() and free_8_pageframes()
in the pageframe allocator and re-writes kbasic_virtual_alloc()
and kbasic_virtual_free() to use it. Now those functions will never
allocate pageframes one by one. They support only 32 and 8 as page count.
If a physically contiguous block of memory is not found, they'll fail.
The only exception is the case of 32 pages: in case of a failure, an
attempt will be made to allocate the 32 pageframes in groups of 8.
Since kbasic_virtual_free() cannot know if the 32 pageframes are
contiguous or not, it always frees them in groups of 8. This seems the
best trade-off between the flexibility of the single pageframe allocation
and the performance of the always-contiguous allocations.