[kmalloc] bool has_some_free_space => full
By changing the name and the meaning of that bool in the block_node
bitfield, the value of new_node is just 8 zeros. This allows the function
evenually_allocate_page_for_node() to use bzero() instead of memset().
Also, bzero() has been optimized a for the case of len divisible by 4.