[bootloader] Make stage3_entry.S to contain both _start and main
For some strange reason, even if stage3_entry.S was passed to the linker
before anything else, the linker decided to put the C function main() at
the very beginning of .text. In the linker script its explicitly stated
that the entry symbol is '_start', but that was not enough. I haven't found
yet a way to force symbols order using the linker script. Therefore, the
hacky workaround is just put both _start and main in stage3_entry.S and
just call the C main function bootloader_main.