[pageframe_allocator] Make kernel to set reserved MBs at runtime
Since it is not very flexible to have the pageframe allocator to have
at compile time the amounts of initial physical memory reserved, the func
mark_pageframes_as_reserved() has been introduced. It is called in main()
after the initialization of the pageframe allocator. This allows the
ultimate flexibility, necessary in order to support for example any
arbitrary physical address for the RAMDISK via multiboot [not implemented
yet]. Implementing this change, required the introduction of a dedicated
INVALID_PADDR == -1, since, at least in the unit tests, now 0 is a valid
return value for the alloc functions.