[kmalloc] Add a linear-mapping region in the kernel virtual space
This change makes the first 256 MB of the physical memory to be linearly
mapped at +KERNEL_BASE_VA (0xC0000000). This makes the allocator much
faster (as the new perf tests prove), allows to switch between paddrs
and vaddr by just using an offset and finally allows the removal of
the 'paddr' field in page_directory_t.