[kmalloc] Remove logic handling allocation for metadata nodes
The allocation for metadata nodes logic was necessary to allow even the
pages containing the metadata nodes (1 per block) to not be necessarily
mapped. The function evenually_allocate_page_for_node() had the essencial
role to allocate some pages for a given node and its neighbour nodes.
In particular, since the size of one node is just 1 byte while the that
function used to map 32KB at a time, that allowed the allocation for
metadata nodes to be a rare thing.

Now that linear mapping is used, keeping the logic for allocation of
metadata nodes seems not really necessary. Kmalloc will support in the
future different heaps, for sure also outside the linear mapping region,
but, it seems inefficient to burn cycles for checking, for each node,
if it has an actual mapping. It will be more convenient to just pre-map
the whole area containing the metadata nodes when the heap is created.