[kmalloc] Improve SIMULATED_CALL()
This changes makes internal_kmalloc() simpler by making SIMULATED_CALL()
to automatically generate a label and store its address in alloc_stack.
This way, in case the simulated call failed the control flow continues
at the caller site as if the simulated call was a real one. No need anymore
for a explicit and ugly labels like "after_left_call".