[entry] Make entry.S much simpler
First, 4-MB pages are used: this way is much simpler to get the same
result as before. No need to create a page table in addition to the
page directory. Second, this simpler code can be written by avoiding
to use the registers EAX and EBX which might contain multiboot info.