[build] Move *run*qemu* scripts in build/
After adding the scripts (with and w/o KVM) to run QEMU with multiboot,
the amount of scripts in project's main directory for running QEMU,
become too large. In addition, the fact that all of those scripts were in
the main directory and referred directly to the build directory as
'./build' was inconvenient for the use-case of multiple build directories
(e.g. debug and release). With this change, this problem completely
disappears.