[fat32] Remove a wrong ASSERT from the FAT32 code
The ASSERT(to_read == d->cluster_size) made sense only in the original
fat_read_whole_file() function, which source has been used to write
fat_read(). After the introduction of fseek(), clearly that ASSERT could
not be kept in fat_read().

BONUS: this change fixes another issue as well: the failure of the fseek
unit test with a "buffers differ" error, in the case of read 0: in this
case, the issue is simply the fact that the buffers (on stack) contain
garbage since fseek returned 0. The fix is simply to zero the buffers
before reading.