[FS] Fix problems caused by use of stdio funcs in init
When stdio funcs are used, dietlibc's printf() doesn't just do a write(),
but it checks with ioctl() if the current output is a tty. Since exOS
didn't have ioctl() yet, the check failed and printf() assumed that there
is no need to write line-by-line the output. Manually forcing the flush
with fflush(stdout) works, but it should work also without it, as it does
on Linux. Therefore, a specially crafted impl of ioctl() has been added
in order to trick dietlibc.