[term] Make term to use video_interface instead of hard-coded func calls
With this change, the term code becomes generic and does not depend on
how actually the characters are displayed on-screen. For the moment,
still the only supported mode is the standard PC text mode (80x25), but
this generalization will allows easily term to work with any other mode
implementing the video_interface. For example, exOS may support a running
its console in a graphical mode, with a framebuffer info passed by the
bootloader. In that case, after implementing the video_interface functions
to work with framebuffer, term will use them seamlessly.