[kcond] Add support for timeout_ticks in kcond_wait()
This super-useful feature allows kcond_wait() to timeout after some time,
if its caller wants that. This feature is particularly useful for an
use-case like the tasklet runner kthread: since it's signalled without
any lock, it can happen that it may loose a signal, while running a
tasklet. Thanks to the timeout_ticks, loosing a signal won't be a big deal.