[process] Implement (partially) waitpid() properly
The previous impl. of waitpid() was just a mini-demo to test that a user
task can be preempted while running in kernel and that the resume after
that will work correctly. Now that we're sure that works, it didn't make
any more sense to keep that impl. Therefore, a proper notification
mechanism has been used to achieve that.