[build] Make scripts/cmake_run to work with any build directory
Until now, scripts/cmake_run was usable only in the project's main dir and
it prepared the build to be in MAIN_DIR/build. Now, it's possible to create
a build directory anywhere and just run MAIN_DIR/scripts/cmake_run there.
The script will use that directory as a build directory and will also copy
there all of the handy scripts for running qemu.