[fs] Add dup to filesystem and use it sys_fork()
The addition of a dup(fs_handle) function was essential in order to allow
a parent to keep its file handles after a child dies (the kernel closes
its file handles).

NOTE: the memory poisoning by kfree() has been commented because it was
making the kmalloc unit-tests to fail. The failure will be investigated.