[kmalloc] Rename kfree() to kfree2()
Now kfree can work both with and without a size argument. The paramater
can be completely removed, but it also makes some sense to have two
versions of kfree: with and without the parameter size. The call-sites
that know at compile-time the size of block that want to free can easily
use kfree2 and a slightly faster kfree. The other sites, will just call
the 1-parameter kfree() wrapper. In the next changes, this wrapper will
be introduced.