[kmalloc] Replace the kmalloc_mutex with disable/enable preemption
If even, theoretically, it is better for the allocator to have its own
lock, that complicates the things since it does not allow (safely)
non-preemptable code to call kmalloc/kfree. What could happen in that case
is that the non-preemptable code might not be able to grab the kmalloc lock
beacuse other preemptable code holds it. This will end up in an ASSERT
since the kmutex code won't be able to put succeed in calling kernel_yield.
For a simple project like exOS, it is acceptable such a commonly used
thing like the allocator to be called inside non-preemptable code. Also,
considering that a kmalloc/kfree pair costs about 500-600 cycles that
sounds not a big deal.