[process32, entry] Re-design the context switch for kernel threads
Currently, kernel threads and the kernel part of the user threads (tasks)
had a dedicated copy of the 'regs' struct in task_info. The save of kernel
context and the switch logic was similar to one used for user tasks, except
for some trickery necessary to handle the fact that in case of no privilege
change, x86 CPUs don't push esp and ss on the stack. The dedicated copy of
'regs' for the kernel state was first pushed on stack by asm_int_handler
or by kernel_yield and then stored in current->kernel_state_regs by either
save_current_task_state() [preemption] or save_current_kernel_task_state()
[yield]. The stack used by the int handler/yield to save the registers was
discarded. Now instead, current->kernel_state_regs has become a pointer to
a 'regs' struct pointing exactly to that part of the stack. This way, the
context switch among kernel threads is faster and we save some memory too.