[process32, entry] Make again the panic being able to show the right regs
After the recent changes in the kernel tasks context switch code, the
panic_save_current_state() got broken, because saving just the esp does
not work, if the panic save function has to return and store the esp.
Therefore, user's state_regs has been used as a store in this unfortunate
corner case.