[printk] Fix a bug in printk_flush() by adding minibuf
The problem: with the previous impl. we first claim in the do-while loop
that we have read the data from the ringbuf and then we actually read it
(while writing to term). The problem with that approach is that if another
printk occurs (3rd nested call), it will find the used field = 0 and so
will use up to sizeof(printk_rbuf) for appending its data, corrupting this
way that the 2nd nested call was about to read. The solution is to use a
minibuf in the do-while loop and, each time, copy the data safely on the
stack. In the worst case when the loop has to iterate more then once, we'll
just have to re-read the data.

BONUS: this commit also adds term_write_char2(c, color) to term.h allowing
its callers to write a char using a specific color without changing the
current color.