[main] Make the kernel to support a custom command following the -cmd opt
When the kernel is booted with multiboot, it is possible to pass options
to through the 'cmdline' multiboot parameter. In case in the cmdline we
have the option '-cmd', the arguments following will be treated as:
<full path of an executable> [arg1 [arg2 [arg3 ...]]]
This allows the kernel to execve() something else instead of /sbin/init,
or to execve() init, but with some additional parameters. Now that init
supports the '--' option allowing to forward the following arguments to
the shell, with this patch it is possible to from the command line to
forward arguments to our simple shell. This will be used by the system
tests.