[shell] Add the option -c making the shell to execute a built-in command
This will be used by the system tests. Now that it is possible through
the qemu command line, through multiboot to pass arguments to the kernel
and in particular execute init with custom arguments, it is possible use
its '--' option to forward options to the shell and therefore ask it,
with '-c <built-in command>' to execute a built-in command and exit.