[term] Introduce and use term actions
This change makes term interface functions to use action elems describing
which action has to be performed and a term_execute_action() function,
instead of directly performing any given action. This is clearly a
preparation change. What will follow is a series of changes making the
term interface functions to decide if any given action can be performed
directly or has to be enqueued in a ring buffer, exactly as it happens
in printk. The whole point is to make term run without disabling the
interrupts and remain safely consistent. In case an interrupt handler
calls a term function, interrupting a term function already running,
its action will be enquequed an executed at the end of the original
term function. Again, exactly like in the printk case.