[self_tests] Fix fault_resumable_perf
The bug was the call to enable_preemption() hitting the oldval > 0 asset.
This happened because now self tests are run always in a dedicated kernel
thread and preemption is enabled by default. Therefore, as expected,
enable_preemption() complained about that. The test has just not been
updated since the change. The fix simply removes the now useless calls
to enable/disable preemption.