[bootloader][main] Use multiboot and make the kernel to depend on that
This change makes the legacy bootloaer to use the multiboot v1 protocol
to boot the kerne and, at the same time, at the kernel side any support
for non-multiboot is removed.

BONUS: term can be initialized also in panic scenarios and even if kmalloc
does not work, thanks to a BSS failsafe buffer.