[bootloader] Remove the flat binary stage3 and use only elf_stage3
By just skipping the first 4 KB, it is possible for the asm bootloader
to jump into an ELF binary, as if it was a flat binary, IF and ONLY IF,
for each section, the file offset and the vaddr differ always by the
same, constant, offset. That's the case for elf_stage3. Using an ELF
binary allows correct debugging in any case. The only price to pay here
for using an ELF binary is that we'll never be able to embed the early
boot code into the binary. That could happen only if a flat binary is
used. But that is too bad for debugging. Creating a parallel ELF binary
of a flat binary just for debugging does not completely work because
addresses in .data and in .bss differ.