[fb_console][panic] Support panic at the earliest possible
Now it's possible, even with the framebuffer console, to see a panic
message very early: immediately after read_multiboot_info(), where we
get the address of the framebuffer. In text mode it's possible instead
to see a panic even before the first call in kmain() because the address
of the character framebuffer is fixed.