[fb_console] Performance improvements
Restore the double buffering but add a 'flush' parameter to vi->set_row():
it allows term to use it when scrolling. In fb_console, the text in first
drawn on the shadow buffer and immediately (if flush is true) flushed.
This approach is ultimately much faster on real hardware, compared to the
draw line-by-line and a big final flush, because of the time-locality.
The scroll_one_line_up() is about 13% faster.