[build][build_toolchain] SWITCH to a LIBMUSL gcc toolchain
This major-impact change makes the whole project to use a musl-based i686
toolchain instead of a "generic" glibc one + many compiler & linker options
necessary to build & link with dietlibc. Libmusl is by far more supported
as a lightweight "libc" than dietlibc: major projectes like "busybox" can
be built with it. Therefore, the #1 reason for switching to this library is
being able to compile busybox and other open source projects. The #2 reason
is using a pre-built gcc toolchain configured to work with libmusl by
default: this makes the build of exOS's userapps simpler but also allows
any project potentially to be built with that toolchain. With dietlibc that
is by far harder and, given the fact that the library is not supported
anymore, does not make much sense to invest any effort in it.

However, it worth to mention that dietlibc helped A LOT in the development
of exOS, because of its very limited requirements towards the kernel.
Supporting libmusl it's possible only today after implementing (partially
or completely) many "newer" syscalls, that were necessary even for a
program to reach main().