[kmalloc] Fix a NASTY bug in set_free_uplevels()
The function was indeed doing correctly its job in coaleshing as much as
possible nodes above the just-freed one BUT, after doing that it did NOT
mark as NOT FULL the nodes above the last coaleshed one. Even if they
cannot be coaleshed, because they are split and half of them is not empty,
still they are clearly NOT FULL. This is a particularly corner case that
was not handled before.