[switch_to_task] Clear timer IRQ mask if !KERNEL_TRACK_NESTED_INTERRUPTS
In commit 9a37d46 that statement has been removed from the function, but
at that time KERNEL_TRACK_NESTED_INTERRUPTS was enabled by default and in
that case, it was correct to remove that because the timer IRQ was never
masked (allowing nested IRQ). After that, KERNEL_TRACK_NESTED_INTERRUPTS
has been make a configurable define and that introduced the bug currectly
being fixed, since both the scenarios are supported.