[exos] Don't use -nostdinc anymore
After implementing "many" syscalls, it turned out that it gets harder and
harder to keep the project completely isolated from the system headers.
In particular, the following problem turned out to be critical: the noarch
part of the kernel needs to build for x86_64 (for the unit tests) and, at
the same time, it needs to put data structures like "struct stat" in common
headers. Those common exos headers get included indirectly by the unit
tests and clearly the exos definitions conflict with the system ones,
visible by the unit tests. In particular, the definitions, even if
isolated (unit tests to never include exos headers) they would be based on
the assumptions that the system is 32-bit (because of the length of some
fields). That may cause nasty and hard-to-debug issues. Therefore, at this
point using linux's system headers seems more convenient because:

	- Saves exOS from copying (carefully!) a lot of stuff from the
	  system headers (less code, less errors).

	- Makes automatically exOS to use the right data structures, when
	  the arch changes.

Clearly, there are still issues with this approach as well. For example,
linux supports the same data structure in different sizes, depending on
the syscall (sys_stat64 vs sys_oldstat). In that case, in order to
implement sys_oldstat(), exOS will still need to have a dedicated
"struct oldstat" which differs from what usermode apps compiled with the
exOS toolchain (modern gcc + libmusl) see as "struct stat". But that still
seems acceptable, considering that, for a pretty long time exOS will run
only programs compiled with its toolchain. And, even when exOS will have
to support "struct oldstat", it still would mean to have just 1 copied
struct instead of 2 (the modern one is given by the system headers).