[exvfs] Change the interface of several exvfs_* funcs
This change makes exvfs funcs to never return directly a handle, but
to use an out-paramater instead. This allows the use of return codes
for handling several errors, typically -ENOMEM. A particularly good
example is open(): it could fail because of the file does not exist or
because there is no enough memory to allocate the file handle.