[exvfs] Fix a bug in check_mountpoint_match() and exvfs_open()
The code was matching a path like "/dev" to "/", even when a mountpoint
"/dev/" existed. That's because mountpoints always end with "/", while
paths never do. Handling for this special case has been added. Now, if
a mountpoint "/xxx/" exists, the path directory "/xxx" refers to the root
("/") of filesystem mounted at "xxx" instead of the directory "xxx"
rooted at "/".

BONUS: find has been added in busybox.