[elf_loader] Fix a bug in the ELF loader
In case two LOAD phdr overlap on the same page and that page is RO, we'll
get a PAGE FAULT with the previous implementation because load_phdr()
was setting pages to RO (when necessary) at its end (so the next
load_phdr(), if a page overlaps, generates a fault). The simple fix is
to do the access setting only after loading all the phdrs.