[interrupts] DEBUG_check_preemption_enabled_for_usermode only for syscalls
After making possible IRQ handlers to ask for a thread to be immediately
scheduled, we created a new state: no preemption + 0 nested interrupts.
When an IRQ comes in this state, the check_preemption_enabled_for_usermode
called by irq_entry() fails causing a kernel panic. A good enough solution
is just to run the check for syscalls. Afterall, it is only a debug check
useful to prevent buggy kernel changes that allow returning to usermode
with preemption disabled: it does not need to run all the time. Check that
at each syscall is good enough.