[process][sched] Fix a nasty bug related with the size of task_info
Brief root-cause. The formula:

	ti = ((task_info *)pi) - 1

Does not work always in release builds when 'pi' is the kernel process.
That's because in create_kernel_process() the 'ti' and 'pi' for kernel
were created this way:

 void create_kernel_process(void)
 {
    static task_info s_kernel_ti;
    static process_info s_kernel_pi;

    [...]

 }

The compiler did some optimization trick and s_kernel_pi was not
immediately after s_kernel_ti for task_info bigger than a certain value.
The solution? Be more explicit and don't allow the compiler to do use
tricks:

void create_kernel_process(void)
{
   static char kernel_proc_buf[sizeof(process_info) + sizeof(task_info)];

   process_info *s_kernel_pi = (process_info *)kernel_proc_buf;
   task_info *s_kernel_ti = (task_info *)(s_kernel_pi + 1);

   [...]
}

That's all.