[fb_console] Remove shadow buffer support
It turned out that PAT can override MTRRs and mark framebuffer's memory as
WC (write-combining) even if there an MTRR marking the same memory region
as UC (uncacheable). With the framebuffer using WC memory, the raw-write
performance increased by a lot (more than 200x). Therefore, now there is
no point in using a double-buffer which, the framebuffer's memory marked
as WC, makes the performance worse instead of improving it.
With this patch, the fb_console is both simpler and much faster.