[build] Introduce the tilck_add_extra_app() macro
This macro makes the integration with extra apps much easier.
It's enough for the external app to just add a call to tilck_add_extra_app()
with the full-path of a binary in order to that deliverable to become a file
dependency of Tilck's main target (the image) and get that deliverable copied
in /usr/bin (in Tilck's fatpart).