[atomics] Implement some (trivial) atomic funcs for GCC <= 4.8
Because of Travis' old compiler (GCC 4.8), that we use to build the 64-bit
kernel_noarch_static_for_test target, does not support C11's atomics
we have to implement them somehow, EVEN IF the implementation is not
efficient. On all the other environments, we use a newer GCC so that the
unit tests can actually test kernel's code with the proper C11 atomics.

NOTE: given the #if used, in NO CASE the real kernel code (not for testing)
will be using the new "trivial" atomic functions defined in atomics.h,
because the compilation will simply fail, if <stdatomic.h> does not exist.