[timer][syscc] Make the 64-bit integer __ticks non-atomic
Having __ticks as 64-bit atomic working on i686 out-of-the-box was caused
by some "magic" existing in GCC 6 (and maybe GCC 5) which used cmpxchg8 in
a loop in order to implement atomic_load, atomic_store, atomic_fetch_add
etc. for 64-bit integers on i686 and for 128-bit integers on x86_64. The
magic has been removed from the code the compiler could emit directly and
moved to libatomic because it was disleading: cmpxchg in a loop is NOT the
same as having HW support for atomic_load, because of the potential
contention and because for read-only memory it was not usable at all.

For the whole story, see the commit:
"Fix __atomic to not implement atomic loads with CAS" by Torvald Riegel.
Patchwork link: https://patchwork.ozlabs.org/patch/721686/

The assumption that 64-bit atomics just work on i686 in Tilck broke the
support for USE_SYSCC=1 when GCC 7 (or newer) was used. In order to fix the
issue, __ticks has been restored to be just non-atomic and get_ticks() uses
cli/sti in order to get the value atomically.

BONUS:
 - The static assert (ATOMIC_LLONG_LOCK_FREE == 2) has been removed.
 - TIME_SLOT_TICKS has been moved to the new sched.h header.