[kgcov][selftests] Fixing synchronization of the self tests
Since now self tests cannot just decide when the test is finished and it's
OK to just turn off the qemu VM, we have to guarantee that in all the pass
cases they will return the control to the caller when all of the async
tasks they've generated are completed. ONLY then, sys_tilck_run_selftest()
can return the devshell can fetch from the kernel the coverage data. This
constraint makes the self tests harder to write but, at the same time, it
makes them better: afterall, if nothing gone wrong (no crash/panic/assert)
there should be NO REASON why the kernel couldn't continue to work as
before.