[sync][process] Remove all the (now) un-necessary no-interrupts blocks
Now that tasklets do NOT use kcond, nor their runners actually affect in
any way the state lists, it is absolutely safe to walk through the state
lists w/o disabling the interrupts. 'kcond' is now (as 'kmutex') completely
disallowed in IRQ context. The only thing that an IRQ handler (!= timer)
can do is to just enqueue a tasklet in a runner. Inside an individual
tasklet, everything can be done, including sleeping on locks, of course.