[list] Introduce a separate type (list) for list heads
It was nice trying to use just a single type both for the list head (which
does not belong to any elem) and for the each list node. Currently, some
structs like task_info belong to some many lists (as a node) that creating
a list (head) for each task will create a lot of confusion: it will look
as just another node. Relying only on variable's name for that is too risky.
The solution is just to "artificially" create a "list" type for each list
head. This way, there will be a *net* distinction between lists and list
nodes.