[build] Build everything with -fno-strict-aliasing
Allowing the (default) strict aliasing rules, especially for low-level C
code like in a kernel is too dangerous. It is typical in kernel code to
use aliased pointers and the strict aliasing rule allows the compiler to
generate "undefined behavior" in such cases. One example: the latest changes
in list.h (introducing struct list) intentionally alias struct list with
struct list_node pointers and that made a simple pointer-comparison to
return 0 instead of 1 in the unit tests. Using an ad-hoc union will solve
the problem, but considering that, in general the kernel in many other
places do use aliased pointers, the safest solution is just to disable this
risky optimization, like the Linux kernel did since the strict aliasing
existed.