[x86_strings] Fix a bug in strlen()
It turned out that using a plain asm() was too unsafe, even if, in theory,
the compiler had all the information needed to generate correct code, in
my opinion. The bug could be reproduced by just using 'asm' instead of
'asmVolatile' in strlen() and observing how, in release builds, the function
sys_tilck_run_selftest() behaves in a weird way. The alternative fix was to
introduce a 2nd unused variable and store EAX there. The compiler generated
correct code, but using 2 instructions more compared to just using
asmVolatile().