[init][system_tests] Make init to wait all the tasks before exiting
Unfortuately, this works correctly only the system tests setup. If we run
the 'waitpid4' test using the system tests infrastructure we get a general
protection fault. The reason for this GPF have to investigated. In the
meanwhile, just saving the work so far with a simple hack that prevents
tests to fail: exit(0) in /bin/init after the shell exits.

BONUS: make 'init' to accept a full command line after '--' instead of just
the arguments for the shell. This allows 'init' to be tested on Linux by
providing the full path to 'devshell', instead of leaving 'init' to assume
that it must be in /bin/devshell.