[debug] Remove loop in debug_qemu_turn_off_machine() as irrelevant
It turned out that the theory in the comments is just wrong. The problem
observed seems to be connected with selftest_fault_resumable_perf_short().
Apparently, after this self test, there is a certain chance that the shell
process never gets the control back. The problem will be further investigated
after collecting more data.