[term][build][debug] Remove TERM_PERF_METRICS
Given the proven speed of Tilck's console implementation, both in text mode
and in framebuffer mode, the time to remove the cludder introduced by the
term perf metrics has come. Such metric not only now is just making term's
code harder to read, but making it available consumes a previous slot in
the Ctrl+Fn debug functions. Therefore, it seems reasonable now to remove
it, hoping that there won't be any need to restore it back in the future.