[elf] Fix a subtle error-case bug in load_elf_program()
In the code allocating the user stack pages the `rc` of the variable was
shadowed by a local `rc` variable. As a result of that, in the case the
loop ended because the kmalloc() or map_page() failed, the outer `rc`
would remain untouched. Of course that lead the user program to crash just
a little later.