[signal] Hack: treat custom handlers as SIG_IGN but keep their value
This hack is a better version of the previous one for one simple reason:
when process forks, custom signal handlers are replaced by SIG_DFL in the
child, while signals ignored with SIG_IGN remain that way. The previous
hack was replacing any custom handler with SIG_IGN causing an incorrect
behavior in the child (keep ignoring the signal instead of performing the
default action). Now, the custom handlers are just ignored in send_signal()
but their function pointer is kept.