[process32] Call arch_specific_free_task() in create_usermode_task()
This change fixes another issue related with TLS support: when execve() is
called, all the GDT and LDT entries HAVE TO be cleared. This was NOT done
before, but everything "seemed to work" just because in no case all per-process
GDT entries have been. Now with ASH that happened. Therefore, when ASH tried
to execute any process, the TLS initialization failed with set_thread_area()
returning -ESRCH.

With this change, the problem has been fixed and now ASH is finally able to
run other programs.