[main] Create the first process (pid 1) in SLEEPING state
This allows us to make init_drivers() run with preemption enabled as in the
past, but in a much safer way because no user process will be able preempt
whatever the kernel does in a sync or async way in init_drivers().

Why init_drivers() cannot be run directly in the kmain?
--------------------------------------------------------

Running init_drivers() in a kernel thread instead of trying to run it from
kmain() is much more flexible: it allows interrupts to work and multiple
kernel threads to run. It happened to be necessary in some corner cases
during the initialization of the PS/2 controller and will be certainly
mandatory in order to initialize more complex stuff, like the ACPI
subsystem.