[signal][process] Make send_signal() to run with preemption disabled
This is a mandatory change: for the first time, a process might affect
another process life's, by sending a signal to it. The problem is: if
preemption is enabled after we got the task_info* pointer, we risk a race
condition in the case the process died in the meanwhile. Therefore, we
the function terminate_process() must require the preemption to be disabled
and send_signal() must accept an TID instead of a pointer, in order to safely
disable the preemption first and then acquire the task_info* pointer.