[process][sched] Make `__current` NON-atomic
The idea of using `__current` as an atomic (relaxed) variable turned out to
be just overly cautious as it is NOT necessary even for architectures which
require more than one instruction in order to load/store a pointer-size
integer. The reasons for that are explained in a comment inside
get_curr_task().