[cmake] Support a permissive mode for env cache str variables
Unfortunately, even if we pass always the same CC when calling CMake,
CMake will sometimes transform its path, using an unclear set of rules.
What is evident is that there no guarantee that even if a canonical path to
a compiler passed as CC will be kept as it is. Therefore, the checking done
by the define_env_cache_str_var() macro might fail. One workaround is to
just stop setting the env variable after the first CMake invokation. The
other one, instead, is to set the PERMISSIVE env variable to 1 and skip
the check.