[process][sched] Make get_task() callable only with preemption disabled
In general, it is UNSAFE to get a task_info* pointer for a task while
preemption is enabled because the task might die in the meanwhile the
pointer is used. No ref-count is used at the moment. Therefore, the simplest
thing to do in order to keep the code safe is to force preemption to be
disabled while get_task() is called and, hopefully, during the whole time
the pointer returned by it is used.