[fb_console] Fix a long-lasting bug with cursor movement
Description: when scrolling, the `term` layer moves the cursor out-of-screen
in order to hide it (because it's the most portable way to do that). But, the
fb_console did not handle that correctly, laking out-of-bounds checks for the
cursor position. This causes Tilck to crash with `termtest -p`, but ONLY for
specific screen resolutions (that made the bug long-lasting).

This patch just adds out-of-bounds checking for cursor's position and solves
the problem.