[clang-format] Add .clang-format to project's root
This clang-format YML config file does NOT define precisely Tilck's coding
style, but makes the tool to format any C file using a coding style as close
as possible to the one used in Tilck's source. Maybe one day the tool will be
powerful enough to support every style. In the meanwhile, the rules in this
file can be used as a rough guideline.