[vfs][fat32][devfs][ramfs] Return -ENOTDIR when appropriate
Until now, Tilck was a little too permissive and it allowed user code to
refer to a file, let's call it 'f1', both as 'f1' and as 'f1/' because it
always used to strip any trailing '/'. Now, the trailing slashes are not
stripped anymore and the specific -ENOTDIR error is returned by open() in
case the entry do exist, but it is NOT a directory. In case the entry does
not exist, -ENOENT is returned, of course.

The handling for this specific case was been implemented in all of Tilck's
file systems: fat32, devfs and the new-born ramfs.